requirements.txt
torch
torchvision
torchaudio
surya-ocr
transformers==4.36.2
Pillow